+++ CHANGELOG.md
@@ -9,6 +9,7 @@ alpha-4
 - [node] copy,move and clone - Target paths automatically append basename if target is a node.
 - [query] Always show path next to resultset
 - [node|shell] Most commands which accept a node path can also accept a UUID
+- [node] `node:list`: Show node primary item value
 
 ### Bugs Fixes